New SelectionGui option
This is a new gui type that takes a list of objects and displays them all on a simple gui. When clicked on it will activate a callback defined in the constructor.  It also involves added SelectionAdaptors which take an object and turn it into an element. These can be defined globally from the SelectionAdaptors class or defined locally and passed to the constructor.